Remove unneeded toInt() call

+++ conclave-enclave/src/main/kotlin/com/r3/conclave/enclave/internal/GramineEntryPoint.kt
@@ -35,7 +35,7 @@ object GramineEntryPoint {
             exitProcess(EXIT_ERR)
         }
 
-        return port.toInt()
+        return port
     }
 
     private fun initialiseEnclave(enclaveClassName: String, hostInterface: SocketEnclaveHostInterface) {
